+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Android: `DecoderConfig` as part of `PlaybackConfig.decoderConfig` to select the preferred decoder.
 
+### Changed
+
+- Update IMA SDK dependency on Android to `3.35.1`
+
 ## [0.40.0] - 2025-03-20
 
 ### Changed